requirements.txt
paho-mqtt==1.6.1
pyzmq==26.0.3
numpy==1.26.4
matplotlib==3.8.4
